Retry pooled connections in case of BadStatusLine
Sometimes, when Apache is used as proxy, HTTP get response may raise a
BadStatusLine exception. This race condition occurs when the Apache
proxy module is being used with HTTP keepalive requests. The backend
server may close the keepalive connection right after httpd has checked
the state of this TCP connection, so httpd sends the request and waits
for a response on a connection that is in fact dead.

Solve the above error by re-sending the request when receiving a
BadStatusLine exception with empty line ('').